Moved countdown logic from App to countdown component
