+++ src/components/CountDownComponent/CountDown/styles.scss
