+++ src/components/CountDownComponent/CountDownSpeedControl/index.jsx
