+++ src/components/CountDownComponent/CountDownSpeedControl/index.test.js
