+++ src/components/CountDownComponent/CountDownSpeedControl/styles.scss
