+++ src/components/CountDownComponent/CountDownStatus/styles.scss
